Split out nodes and thumnail handling into dedicated modules
